backend/requirements.txt
beautifulsoup4==4.13.4
Django==4.2.9
pdfplumber==0.7.0
mongoengine==0.27.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
requests==2.32.3
django-cors-headers>=4.0.0,<5.0.0
djangorestframework==3.14.0
pymupdf==1.21.0
gunicorn==20.1.0
pymongo==4.5.0